Organised most files with new code and comments
