+++ src/components/Reloader.jsx
@@ -9,9 +9,7 @@
 import React from 'react';
 
 import lottie from 'lottie-web';
-import reversedProgressBar from './animation/files/reversed-progress-bar.json';
-
-import settings from '../settings/general';
+import reversedProgressBar from '../media/animations/reversed-progress-bar.json';
 
 /* * */
 /* * * * */
@@ -43,7 +41,7 @@ export default class Reloader extends React.Component {
     // Set Animation speed
     // Animation is 1 second by default,
     // so total duration is 1 divided by desired speed.
-    const defaultSpeed = settings['max-inactive-seconds-allowed'];
+    const defaultSpeed = 20; // seconds
     const desiredSpeed = defaultSpeed / (this.props.speed || 1);
     this.animation.setSpeed(1 / desiredSpeed);
